requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
fastapi==0.128.5
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
click==8.3.1
pyee==13.0.0
greenlet==3.3.1
playwright==1.58.0
python-dotenv==1.0.1
psycopg2-binary==2.9.11
SQLAlchemy==2.0.48